proto_tree_add_uint_format_value() puts the field name into the textual representation for you; the format string should not include the field name.
Add protocol items to the SliMP3 tree, not to the top-level tree.

If we're fetching a 16-bit value from the protocol tree and multiplying
it by 2, it won't necessarily fit in a guint16; make the variables used
for that guints.

The sequence field of the MP3 data ack packet isn't being fetched from
the packet, so just use proto_tree_add_item() for it (rather than using
a value fetched for a previous field).

Use %u to format unsigned values.

svn path=/trunk/; revision=48028