Give 3GPP TS numbers for GTP v0, GTP v1, and GTP'.
Give URLs for the 3gpp.org pages for all those specs.

Add #defines for all the GTP v1 extension header types we handle.

"gtp_prime" is a Boolean; make it a gboolean.

Dissect the first 4 octets of the header one field at a time, so that if
the packet is cut short by a snapshot length we at least dissect what's
there.

32.295 isn't entirely clear on what the 20-byte header for GTP' v0 is;
assume it's the same as the header for GTP v0.

Once we've fetched the length field from the fixed-length portion of the
header, set the length of the tvbuff to the sum of the offset past the
fixed-length portion and the length field, to catch running past that
value.

Use GTP_E_MASK|GTP_S_MASK|GTP_PN_MASK as the mask for testing for the
presence of those fields, to make it a bit clearer what's being checked
for.

Don't actually add those fields to the protocol tree unless the flag for
the field is set.

We only need one chunk of code to handle extension headers.

Make that chunk a loop, and put the header in as an FT_NONE item, with
the length, header data, and next header under it.  Put the initial next
header field in as well.  (We treat this like IPv6 extension headers,
with the next header field being part of the previous header, rather
than like a set of TLVs, with the next header field being the type value
of its header.)

Fail if the extension header length is zero.

Use the reported length when processing IEs or T-PDU payload.

svn path=/trunk/; revision=48068