Fix some variable names to better indicate what they do.
Fix indentation.

Fix a proto_tree_add_uint_format_value() call not to include the name of
the field - proto_tree_add_uint_format_value() will add that for you.

Have dte_address_util() take the offset of the address as an argument;
it's not always at the same offset from the beginning of the facility.

Have it return the pointer to the generated string directly, rather than
through a pointer argument.

Create only one subtree for each facility, and give it a text description
of the facility code rather than the numerical value of the facility
code.  Make the top-level item for the facility cover all the bytes of
the facility, including code, length if present, and parameters.

Dissect the end-to-end transit delay and priorities facilities
completely.  Also, fix an incorrect use of "transmit delay" to say
"transit delay".

Get rid of the last of the spaces preceding colons in "Field: value"
descriptions and in a "default:" case label.

Do the data vs. non-data packet thing ith

	if (PACKET_IS_DATA(pkt_type)) {
		...
	} else {
		...
	}

rather than, in effect, doing the "else" with a break; that makes the
code a bit clearer.

Put the logical channel number into the protocol tree in common code for
the default case, rather than doing it separately for data and non-data
packets.  Clean up the dissection of non-data packets to add entries
before updating the columns, so that we don't throw an exception
updating the columns before we get to add items that wouldn't throw
exceptions.  Clear the Info column early in the dissection, in case we
throw an exception before getting to set it and thus leave behind the
column information for the protocol atop which we're running.

svn path=/trunk/; revision=48093