SPC-3 and later has a 16-bit allocation length for the INQUIRY allocation length.
While SPC-2 only has one byte for allocation length

Change this to treat allocation length as a 16 bit quantity starting one byte prior to the current single byte that wireshark dissects.
This makes it correct for SPC-3 and later  and still works for SPC-2 since that first byte is reserverd, == must be 0, in SPC-2 and prior.
  

svn path=/trunk/; revision=48258