From Sylvain Munaut via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8487
Use enum values instead of manual constants in value_strings. This was
introduced by r47218, which was designed to fix r47209, which would have been
correct in the first place if the value_strings hadn't been full of hacks.

I think everything is correct now. Perhaps it may be worth using extended
value_strings though, since most of the tables seem to be consecutive in order?

svn path=/trunk/; revision=48345