Follow-up to r48395 and bug #8494.
Make libnl a --with flag instead of an --enable flag, since it is an optional
package. Make it a three-way flag, where 'no' disables, 'yes' enables (and
errors if it can't be found), and 'auto' enables (but doesn't complain if it
can't be found). Default to 'auto', which was our old behaviour.

[This is my first real foray into autoconf, so hopefully I got everything right]

svn path=/trunk/; revision=48398