And it happens again...
In Yet Another Protocol, implementation A neglected to set the padding bytes
to 0 and implementation B barfed on said padding (interestingly this
protocol's spec does not include the IETF-normal "receiver MUST ignore the
padding" blurb).

So:

Add the AVP to the dissection tree and add an expert info for when it's not
zero.

Also re-order a few of the hfs and remove a couple unneeded temporary
variables.

svn path=/trunk/; revision=48488