From G.Gerrits via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8501 HomeplugAV dissector: Decoding HPAV-1.1 Tone Map Tx/Rx Characteristics Req/Conf (0xA070/0xA071/0xA090/0xA091)
Dissecting HPAV-1.1 Tone Map Characteristics Req/Conf (0xA070/0xA071) was not correct.
Also found that MMType 0xA070/0xA071 is used for the Tx ToneMaps and the MMType 0xA090/0xA091 for the Rx ToneMaps. Added types and dissector for the Rx MMType. Although the Rx-Conf-Structure is almost identical to the Tx is did duplicate the dissector code to avoid mixups.
Added some additional Infos to the carrier dissector output (Carrier#, #Modulated-Bits, Average#Modulated-Bits/Carrier).

Also extended the homeplug_av_tone_map_vals with 0x08 for 12bit 4096-QAM (the AR74xx should support this kind of modulation) and changed the mask to 0x0f accordingly

svn path=/trunk/; revision=48592