From Tim Ringenbach: makes the TDMoE dissector call the LAPD bitstream dissector on the D-Channel. As a result, LAPD calls Q.931, and you can actually see call setup and tear down, instead of just a hex dump.
It adds a preference for which channel the D-Channel is.

It patches the LAPD code to fix a few bugs, not pass the checksum to Q.931 (who isn't expecting it), to register the lapd-bitstream dissector, and to mark packets with aborts or resets.

also storing more data on the lapd_byte_state_t.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8374

svn path=/trunk/; revision=48743