Don't use ENC_NA for integer fields longer than 1 byte; use ENC_BIG_ENDIAN or ENC_LITTLE_ENDIAN.
Treat the AP PS Buffer State subfield of the QoS field the same way we
treat other subfields in the second byte of the QoS field - show it as
the upper 8 bits of the (shown as big-endian) QoS field.  Fix a bitmap
while we're at it.

Show the channel map as an FT_NONE - it's a structure with two bytes.

svn path=/trunk/; revision=48924