As suggested by Evan:
Point users to https://code.google.com/p/pyreshark/ as an alternative to the
current Python support.

While we're in there, make it harder to enable Python: change it from
"--with-python" to "--with-broken-python" just to prevent people from enabling
unless they really mean it (are going to work on fixing it).

svn path=/trunk/; revision=49138