From Dirk Jagdmann via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8640
Fix dissection of the NTLMSSP v2 Challenge Response. The old code would loop
over the Attributes until the end of the Response, however in reality the last
values of a Response are four 0 bytes and maybe padding. Also change the names
of the values to match those from the Microsoft specification. Also change the
name of one flags bit from unknown to "Anonymous", again according to the
Microsoft spec.

svn path=/trunk/; revision=49258