As suggested by Moshe Kaplan in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8712 :
When pointing people to the mailing lists, point them directly to the lists
page rather than explaining where to look for the link on the main page (the
directions were several years out of date).

While there also fix up some of the indentation and add a link to the Q&A site
in the developer's guide.

svn path=/trunk/; revision=49553