Use older versions of some libraries, and build 32-bit, when building for Leopard.  Fix some issues that came up with the older versions.
We don't support building *on* Leopard - we check for that up front - so
eliminate some code to handle that.

When uninstalling, check for the installed version by looking at the
name of the -done file, so we don't have to trust the _VERSION settings.

We don't appear to need to build our own versions of libpng or pixman
when building *on* Snow Leopard *for* Leopard.

The libffi configure patch for GLib appears not to be necessary -
perhaps building and installing pkg-config *before* building GLib makes
it no longer necessary - and it doesn't work with older versions of
GLib, so don't apply it.

svn path=/trunk/; revision=50456