On OS X, set the rpath for executables to include @executable_path/../lib as well as /usr/local/lib, so we can use @rpath in the install names in the executables and libraries in the application bundle.
Have the osx-app.sh script tweak all references to libraries from
/usr/local/lib in all executables, libraries, and plugins in the app
bundle to use @rpath.  (The "all" is important; it fixes the GTK+ crash
mentioned in the comment in osx-app.sh.  The notion of doing all of them
came from the osx-app.sh script in a newer version of Inkscape.)

This renders the setting of DYLD_LIBRARY_PATH in the wrapper scripts in
the bundle unnecessary; remove it.  (Ideally, we should try to get rid
of the wrapper scripts entirely, but that might have to wait for us to
switch to using Qt.)

svn path=/trunk/; revision=50560