One more try at fixing all the issues reported in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8908 :
Rename check_expand_children() to check_expand_trees() since the function
doesn't just do the children of the current path/iter.

Add a parameter to that function which controls whether the parent (or just
the children) get expanded.

As suggest by Jakub in the bug: block further calls to expand_tree() when
we're expanding all these trees.  Add a comment telling callers of
check_expand_trees() to do that.

svn path=/trunk/; revision=50595