From Michael Mann:
the heuristics of NDMP require 28 bytes to be present in order for it to be dissected as NDMP.  NDMP is purely a heuristic dissector because TCPENCAP shares the same TCP port.
Since NDMP officially registered port 10000 with IANA, allow it to use dissector_add_uint() instead of packet-ipsec-tcp.c, and by default have TCPENCAP only rely on its heuristics.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5462

svn path=/trunk/; revision=51180