From Peter Wu via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9144 [PATCH] Update TLS Ciphers
Really add support for AEAD ciphers (GCM)

GCM uses counter mode with authentication tags (the latter is currently
not supported). As for the key material, there is no MAC (because the
auth tag is supposed to verify the authenticity).

Finally, correct the GCM cipher suite definitions: IV block size of
4 bytes and GCM instead of CBC mode.

svn path=/trunk/; revision=52150