In a Simple Packet Block, the captured length isn't the block length minus the lengths of the two length fields and the packet length field, it's the minimum of that and the packet length, as there might be padding.
Fixes one problem found by the file in bug 9200.

While we're at it, pcapng_read_packet_block() and
pcapng_read_simple_packet_block() return an integer, not a Boolean;
return 0, not FALSE (they have the same value, but returning 0 makes it
clearer that the return value isn't restricted to TRUE or FALSE).

svn path=/trunk/; revision=52241