From Ed Beroset via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9245
Fix test suite so that it doesn't attempt to use color if ncurses is not
installed.  This came up when I attempted to use the test suite under Windows
with cygwin installed but not ncurses.  The result is not entirely broken but
the error messages saying that "tput is not a valid command" make the menus and
results difficult to read.

svn path=/trunk/; revision=52449