Hacky fix for https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9248 Just break out of the loop if offset doesn't go up.
There's almost certainly a better fix - the dissector is weird, and I'm not sure
if all the _length_remaining() checks are important or legacy, and what affect
they have on this issue.

At the very least this will pacify the fuzzbots until somebody has time to
figure it out properly.

svn path=/trunk/; revision=52458