Move the QT_MACEXTRAS_LIB cmakedefine in cmakeconfig.h.in to the equivalent location in config.h.in, to simplify comparing the two files.
Add some tests from the autoconf script to CMake:

	check for setresgid() and setresuid();

	check for struct stat having an st_flags member.

Get rid of INTTYPES_H_DEFINES_FORMATS; we don't appear to check for it,
and we don't use it (we're using GLib's formatting functions, which have
their own way of handling 64-bit integers).

Get rid of STDC_HEADERS; it's in autoconf because autoconf was
originally developed back when you couldn't rely in ANSI C and has never
been removed, and something we're using for other purposes checks for
it, not because we explicitly test for it, but we don't test it in the
code, and CMake doesn't even check for it.

svn path=/trunk/; revision=52522