First cut at having this handle Qt.  Adds a TOOLKIT variable that you set to "qt", "gtk2", or "gtk3".
Work done on a machine with the command-line build tools, but not Full
Frontal Xcode, installed, so it couldn't build Qt, so it hasn't been
tested.  It *has* been fixed to complain in that case, rather than
letting the Qt build fail with a somewhat less useful message talking
about Xcode licensing.

svn path=/trunk/; revision=52776