Finish implementing https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7263 :
Change cflow.sysuptime from an FT_UINT32 (milliseconds since the router booted)
to an FT_RELATIVE_TIME (seconds since the router booted).  I don't imagine
anyone will care if we show them seconds or milliseconds and it satisfy the
user's request (in that bug) to compare cflow.sysuptime to cflow.timeend (which
is already an FT_RELATIVE_TIME).

(If someone does care, we could always display the field twice, once in
milliseconds and once in seconds.)

svn path=/trunk/; revision=52821