From Peter Wu via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9341 :
SVN rev 50525 tried to fix out-of-source-tree builds, but did so by
writing the .rcc.cpp files to the source tree. This was noticed when
running `make distcheck`. This patch reverts that behavior, and creates
the image/ directory to fix out-of-tree builds.

In other words, fix things so "make distcheck" with Qt works in out-of-tree
builds too.

svn path=/trunk/; revision=52926