Make it so out-of-tree builds (autofoo and cmake) have access to the various data files (diameter/*, COPYING, manuf) when running *shark from the build directory.
Do this by passing in the top-level source directory as a compile-time
definition (unfortunately this has to be in the top-level Makefile too because
some programs link directly with epan/filesystem.c).

The plugins dir is no longer below the datafile directory but rather the
progfile directory (if we have one).  Handle the special case of AUTHORS-SHORT
(a data file but a generated one) by checking the file name before building
the path.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5664

svn path=/trunk/; revision=52940