Correctly report segments marked with REASSEMBLE_FLAGS_NO_FRAG_NUMBER.  Bug 9304 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9304)
Instead of incrementing the offset for each new segment by one we add the length of the segment so that each segment is correctly shown in the segment list.

It proves to be very useful to find which packet (segment) is causing an application dissector to go wrong.

From Matthieu Patou

svn path=/trunk/; revision=53118