Have the gsm_rlcmac_ul and gsm_rlcmac_dl use the data parameter of the dissector function instead of pinfo->private_data to receive a RlcMacPrivateData_t structure.
This work was done in bug 7615 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7615), but the tie in to use the RlcMacPrivateData_t structure (passed by the GSMTAP dissector) was never completed.  Whenever that picks up again, the data parameter of the dissector should be used instead of pinfo->private_data.

svn path=/trunk/; revision=53128