Bugfix Negative Fibre Channel scsi_time value.  Bug 3209 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3209)
The solution was to move the itlq_nexus_t structure handling from the FC dissector to the FCS dissector since it's actually the one with the LUN field.  OXID+LUN makes a unique exchange at FC layer so track that separately.  Also needed some additional OXID->LUN mapping at the FC layer to keep LUN correctly persistent.

Not backporting because this was partially made possible by r53569, which won't be backported.

svn path=/trunk/; revision=53620