Add G.722, G.726 and SBC codecs.  G.722 and G.726 are from bug 5619 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5619) and SBC is from bug 7893 (https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7893).
Codecs are added, but (intentionally) not hooked to the RTP player as a "more generic architecture" is desired.  There are some discussions in bug 7893 on how to do this.  One thing to add would be how to handle codecs that may not be supported on all platforms.  Should the codec not be "registered" at all (with a #define over the whole module) or should it's register functions be stubbed (with a #define in each function that requires a non-supported library)

svn path=/trunk/; revision=53676